ajout pour ne pas pouvoir supprimer projet par defaut

+++ src/main/java/com/todoitproject/service/impl/ProjectService.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.todoitproject.dto.DtoBoolean;
 import com.todoitproject.dto.DtoProject;
 import com.todoitproject.dto.DtoProjectDescription;
 import com.todoitproject.dto.DtoProjectName;
@@ -35,6 +36,7 @@ public class ProjectService implements IProjectService {
 
 	/**
 	 * Création d'un nouveau projet
+	 * 
 	 * @param DtoProject
 	 * @return dtoproject
 	 */
@@ -44,7 +46,7 @@ public class ProjectService implements IProjectService {
 		EProject eProject = new EProject();
 		eProject.setNom(dtoproject.getNom());
 		eProject.setDescription(dtoproject.getDescription());
-		
+
 		if (this.checkUserByLog(dtoproject.getId_user())) {
 			eProject.seteUser(this.getUserByLog(dtoproject.getId_user()));
 			projectRepository.save(eProject);
@@ -65,31 +67,31 @@ public class ProjectService implements IProjectService {
 	public List<DtoRProject> listProject(long id) {
 		List<EProject> projets = projectRepository.findByUser(id);
 
-		return projets.stream()
-				.map(a -> new DtoRProject(a))
-				.sorted((b1,b2)-> b1.getNom().compareTo(b2.getNom()))					
+		return projets.stream().map(a -> new DtoRProject(a)).sorted((b1, b2) -> b1.getNom().compareTo(b2.getNom()))
 				.collect(Collectors.toList());
 
 	}
-	
-	/* choisir la bonne methode A NE PAS UTILISER
-	@Override
-	public List<DtoProject> listProjectAll(long id) {
-		List<EProject> projets = projectRepository.findAll();
-		Optional<EProject> optProject = projectRepository.findById(id);
-		
-		return projets.stream().map(a -> new DtoProject(a))
-				//.filter(b->b.getId_user() == (this.getUserByLog(id))) // probleme pas meme type
-					.filter(b->b.setId_user(().getUserByLog(id)this.getUserByLog().get))							
-				.collect(Collectors.toList());
-		  
-	}*/
+
+	/*
+	 * choisir la bonne methode A NE PAS UTILISER
+	 * 
+	 * @Override public List<DtoProject> listProjectAll(long id) { List<EProject>
+	 * projets = projectRepository.findAll(); Optional<EProject> optProject =
+	 * projectRepository.findById(id);
+	 * 
+	 * return projets.stream().map(a -> new DtoProject(a))
+	 * //.filter(b->b.getId_user() == (this.getUserByLog(id))) // probleme pas meme
+	 * type .filter(b->b.setId_user(().getUserByLog(id)this.getUserByLog().get))
+	 * .collect(Collectors.toList());
+	 * 
+	 * }
+	 */
 
 	/**
 	 * Trouver un projet
 	 * 
 	 * @param l'id de l'utilisateur
-	 * @return un projet avec DtoRProject avec son nom et sa description 
+	 * @return un projet avec DtoRProject avec son nom et sa description
 	 */
 	@Override
 	public DtoProject oneProject(long id) {
@@ -109,28 +111,42 @@ public class ProjectService implements IProjectService {
 
 	/**
 	 * Supprimer un projet
+	 * 
 	 * @param l'id du projet
 	 * @return true pour confirmer la suppression du projet
 	 */
 	@Override
-	public boolean deleteProject(long id) {
+	public DtoBoolean deleteProject(long id) {
+		DtoBoolean dtoBoolean = new DtoBoolean();
 		Optional<EProject> oeProject = projectRepository.findById(id);
 
 		if (oeProject.isPresent()) {
 
-			projectRepository.delete(oeProject.get());
-			
+			if (oeProject.get().getId() != 1) {
 
-		} else {
-			throw new NotFoundException("Ce projet n'existe pas");
-			
+				projectRepository.delete(oeProject.get());
+				// System.out.println("testdelete");
+				dtoBoolean.setBool(true);
+
+			} else {
+				// System.out.println("testdeletefaux");
+				dtoBoolean.setBool(false);
+
+			}
 		}
-		return true;
+
+		else {
+			dtoBoolean.setBool(false);
+
+		}
+		return dtoBoolean;
 	}
-	
+
 	/**
 	 * Modifier le nom d'un projet
-	 * @param l'id du projet et le DtoProjectName pour remplacer le champ nom dans entité
+	 * 
+	 * @param l'id du projet et le DtoProjectName pour remplacer le champ nom dans
+	 *        entité
 	 * @return true
 	 */
 	@Override
@@ -146,7 +162,6 @@ public class ProjectService implements IProjectService {
 
 			eProj.setNom(dtoProjectname.getNom());
 			projectRepository.save(eProj);
-		
 
 		} else {
 			throw new NotFoundException("Ce projet n'existe pas");
@@ -157,7 +172,8 @@ public class ProjectService implements IProjectService {
 	/**
 	 * Modifier la description d'un projet
 	 * 
-	 * @param l'id du projet et le DtoProjectDescription pour remplacer le champ description dans l'entité
+	 * @param l'id du projet et le DtoProjectDescription pour remplacer le champ
+	 *        description dans l'entité
 	 * @return true
 	 */
 
@@ -182,27 +198,29 @@ public class ProjectService implements IProjectService {
 		return true;
 	}
 
-/**
- *  Vérifie qu'un utilisateur est bien connecté 	
- * @param id
- * @return
- */
-	
+	/**
+	 * Vérifie qu'un utilisateur est bien connecté
+	 * 
+	 * @param id
+	 * @return
+	 */
+
 	private boolean checkUserByLog(long id) {
-		
+
 		Optional<EUser> oeUser = userRepository.findUserById(id);
-		
+
 		if (oeUser.isPresent()) {
-		
+
 			return true;
 		} else {
-			
+
 			return false;
 		}
 	}
 
 	/**
 	 * Récupère l'id de l'utilisateur connecté
+	 * 
 	 * @param id
 	 * 
 	 */
